+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 * Added support for Cucumber-JVM 1.2.4 and 1.2.5 (#13, #15)
+* Fetch from source repository before publishing to Cucumber Pro (#9)
 
 ### Changed